Cell: move checkbox & dropdown inside Cell folder
