+++ src/components/Cell/CellCheckbox/CellCheckbox.css
