+++ src/components/Cell/CellCheckbox/CellCheckbox.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { Icon24CheckBoxOff, Icon24CheckBoxOn, Icon24CheckCircleOff, Icon24CheckCircleOn } from '@vkontakte/icons';
-import { getClassName } from '../../helpers/getClassName';
-import { usePlatform } from '../../hooks/usePlatform';
-import { classNames } from '../../lib/classNames';
-import { ANDROID } from '../../lib/platform';
-import { CellProps } from '../Cell/Cell';
+import { getClassName } from '../../../helpers/getClassName';
+import { usePlatform } from '../../../hooks/usePlatform';
+import { classNames } from '../../../lib/classNames';
+import { ANDROID } from '../../../lib/platform';
+import { CellProps } from '../Cell';
 import './CellCheckbox.css';
 
 export type CellCheckboxProps = Pick<CellProps, 'defaultChecked' | 'checked'> & React.InputHTMLAttributes<HTMLInputElement>;
